pyproject.toml
[project]
name = "bricksmith"
version = "0.1.0"
description = "POV: deck due tomorrow. You describe, AI draws, judge roasts, DSPy fixes. No more 2am Lucid."
readme = "README.md"
requires-python = ">=3.11"
authors = [
    { name = "David O'Keeffe" }
]
dependencies = [
    "mlflow>=3.8.1",
    "openai>=1.0.0",
    "google-genai>=0.4.0",
    "google-auth>=2.27.0",
    "google-cloud-aiplatform>=1.40.0",
    "pyyaml>=6.0.1",
    "pydantic>=2.5.0",
    "pydantic-settings>=2.1.0",
    "click>=8.1.7",
    "pillow>=10.2.0",
    "rich>=13.7.0",
    "python-dotenv>=1.0.0",
    "pandas>=2.0.0",
    "requests>=2.31.0",
    "dspy-ai>=2.0.0",
    "databricks-sdk>=0.20.0",
    "mcp>=1.0.0",
    "duckdb>=1.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    "black>=24.1.0",
    "ruff>=0.1.0",
    "mypy>=1.8.0",
    "types-pyyaml",
]
web = [
    "fastapi>=0.109.0",
    "uvicorn>=0.27.0",
    "psycopg2-binary>=2.9.9",
]

[project.scripts]
bricksmith = "bricksmith.cli:main"
rfp-refiner = "rfp_refiner.cli:main"
rfp-store = "rfp_refiner.output_store:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/bricksmith", "src/rfp_refiner"]

[tool.pytest.ini_options]
pythonpath = "src"
testpaths = ["tests"]
addopts = "--cov=bricksmith --cov-report=html --cov-report=term-missing"

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"

[tool.ruff.lint.per-file-ignores]
"src/bricksmith/cli.py" = ["E402"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
ignore_missing_imports = true
